Reorganized files to remove `bot` folder.
