+++ src/commands/disable.js
@@ -4,7 +4,7 @@ import {
   PermissionFlagsBits,
 } from 'discord.js';
 import pino from 'pino';
-import instances from '../../model/instances.js';
+import instances from '../model/instances.js';
 
 const logger = pino();
